requirements.txt
fastapi
uvicorn[standard]
gunicorn
torch
torchvision
pillow
python-multipart
pydantic
pydantic-settings
python-dotenv
requests
